fix: 修复ci报错

+++ internal/pkg/image.go
@@ -126,11 +126,18 @@ func DownloadAndSaveImages(imgURL string, quality int) (string, error) {
 	}
 
 	_, err = io.Copy(tempFile, resp.Body)
-	resp.Body.Close()
-	tempFile.Close()
+	errs := resp.Body.Close()
+	if errs != nil {
+		logger.Error(nil, "Close response body error: %s", errs.Error())
+	}
+	errs = tempFile.Close()
+	if errs != nil {
+		logger.Error(nil, "Close temp file error: %s", errs.Error())
+		return "", fmt.Errorf("close temp file error: %w", err)
+	}
 
 	if err != nil {
-		errs := os.Remove(tempFilePath)
+		errs = os.Remove(tempFilePath)
 		if errs != nil {
 			logger.Error(nil, "Remove temp file error: %s", errs.Error())
 		}
@@ -140,7 +147,7 @@ func DownloadAndSaveImages(imgURL string, quality int) (string, error) {
 	// 转换为WebP
 	file, err := os.Open(tempFilePath)
 	if err != nil {
-		errs := os.Remove(tempFilePath)
+		errs = os.Remove(tempFilePath)
 		if errs != nil {
 			logger.Error(nil, "Remove temp file error: %s", errs.Error())
 		}
@@ -148,9 +155,14 @@ func DownloadAndSaveImages(imgURL string, quality int) (string, error) {
 	}
 
 	img, _, err := image.Decode(file)
-	file.Close()
+	errs = file.Close()
+	if errs != nil {
+		logger.Error(nil, "Close temp file error: %s", errs.Error())
+		return "", fmt.Errorf("close temp file error: %w", err)
+	}
+
 	if err != nil {
-		errs := os.Remove(tempFilePath)
+		errs = os.Remove(tempFilePath)
 		if errs != nil {
 			logger.Error(nil, "Remove temp file error: %s", errs.Error())
 		}
@@ -161,7 +173,7 @@ func DownloadAndSaveImages(imgURL string, quality int) (string, error) {
 	webpPath := strings.TrimSuffix(tempFilePath, filepath.Ext(tempFilePath)) + ".webp"
 	outFile, err := os.Create(webpPath)
 	if err != nil {
-		errs := os.Remove(tempFilePath)
+		errs = os.Remove(tempFilePath)
 		if errs != nil {
 			logger.Error(nil, "Remove temp file error: %s", errs.Error())
 		}
@@ -176,16 +188,29 @@ func DownloadAndSaveImages(imgURL string, quality int) (string, error) {
 		Lossless: false,
 		Quality:  float32(quality),
 	})
-	outFile.Close()
+	errs = outFile.Close()
+	if errs != nil {
+		logger.Error(nil, "Close output file error: %s", errs.Error())
+		return "", fmt.Errorf("close output file error: %w", err)
+	}
 
 	if err != nil {
-		os.Remove(tempFilePath)
-		os.Remove(webpPath)
+		errs := os.Remove(tempFilePath)
+		if errs != nil {
+			logger.Error(nil, "Remove temp file error: %s", errs.Error())
+		}
+		errs = os.Remove(webpPath)
+		if errs != nil {
+			logger.Error(nil, "Remove webp file error: %s", errs.Error())
+		}
 		return "", fmt.Errorf("encode webp error: %w", err)
 	}
 
 	// 删除原始临时文件
-	os.Remove(tempFilePath)
+	err = os.Remove(tempFilePath)
+	if err != nil {
+		logger.Error(nil, "Remove temp file error: %s", err.Error())
+	}
 
 	outputPath = webpPath
